requirements.txt
chainlit
python-dotenv
openai-agents
tavily-python